requirements.txt
pandas
openpyxl
python-dotenv
pytesseract
pillow
numpy
pymupdf
pdf2image
msal
requests
